AI seems a bit confused:-
What do you mean by "we can't do that during initialization because gamma is currently at the interrupted level."? It's not clear which "initialization" you're referring to. Can't we simply ensure this "initialization" is done at the beginning of a fade-out? (as the values at this point should always be useable).

Also, what does "Fade-out or fade-in from black" mean? Surely only a fade-in can happen from black, can't it? A fade-out is always going to be some something non-black, right?